+++ cli/scripts/post-build.js
@@ -11,7 +11,6 @@ const npmModulesToCopy = [
   'react',
   'react18',
   'vue',
-  'vue2',
   'angular',
   'angular-signals',
   'svelte',