+++ packages/frontend-shared/src/utils/icons.ts
@@ -9,9 +9,7 @@ import LogoSvelte from '../assets/logos/svelte.svg'
 export const FrameworkBundlerLogos = {
   webpack: LogoWebpack,
   vite: LogoVite,
-  vue2: LogoVue,
   vue3: LogoVue,
-  vueclivue2: LogoVue,
   vueclivue3: LogoVue,
   nextjs: LogoNext,
   react: LogoReact,