+++ packages/scaffold-config/src/dependencies.ts
@@ -7,15 +7,6 @@ export const WIZARD_DEPENDENCY_WEBPACK = {
   minVersion: '^4.0.0 || ^5.0.0',
 } as const
 
-export const WIZARD_DEPENDENCY_VUE_2 = {
-  type: 'vue',
-  name: 'Vue.js 2',
-  package: 'vue',
-  installer: 'vue@2',
-  description: 'The Progressive JavaScript Framework',
-  minVersion: '^2.0.0',
-} as const
-
 export const WIZARD_DEPENDENCY_VUE_3 = {
   type: 'vue',
   name: 'Vue.js 3',
@@ -151,7 +142,6 @@ export const WIZARD_DEPENDENCIES = [
   WIZARD_DEPENDENCY_NEXT,
   WIZARD_DEPENDENCY_REACT,
   WIZARD_DEPENDENCY_REACT_DOM,
-  WIZARD_DEPENDENCY_VUE_2,
   WIZARD_DEPENDENCY_VUE_3,
   WIZARD_DEPENDENCY_ANGULAR_CLI,
   WIZARD_DEPENDENCY_ANGULAR_DEVKIT_BUILD_ANGULAR,
@@ -227,7 +217,6 @@ const testingDependenciesOfInterest = [
   '@playwright/experimental-ct-react',
   '@playwright/experimental-ct-svelte',
   '@playwright/experimental-ct-vue',
-  '@playwright/experimental-ct-vue2',
   '@playwright/experimental-ct-solid',
   '@playwright/experimental-ct-react17',
   'axe-core',